pkg/apparmor: move all linux-code into apparmor_linux*
For easier maintenance and clearer structure of the code.
